- Created the words API

+++ backend/dependencies/groups_db.py
@@ -26,7 +26,13 @@ async def create_new_group(group_name: str, conn: asyncpg.Connection) -> int:
     return group_id
 
 
-async def add_word_to_specific_group(group_id: int, word_id: int, conn: asyncpg.Connection) -> int:
+async def add_word_to_specific_group(group_id: int, word: str, conn: asyncpg.Connection) -> int:
+    word_id = await conn.fetch("SELECT id FROM words WHERE word = $1", word)
+    if not word_id:
+        word_id = await conn.fetchval(
+            'INSERT INTO words (word) VALUES ($1) RETURNING id',
+            word
+        )
     await conn.execute(
         'INSERT INTO word_group_assignments (group_id, word_id) VALUES ($1, $2)',
         group_id, word_id
@@ -34,7 +40,10 @@ async def add_word_to_specific_group(group_id: int, word_id: int, conn: asyncpg.
     return word_id
 
 
-async def remove_word_from_specific_group(group_id: int, word_id: int, conn: asyncpg.Connection) -> bool:
+async def remove_word_from_specific_group(group_id: int, word: str, conn: asyncpg.Connection) -> bool:
+    word_id = await conn.fetch("SELECT id FROM words WHERE word = $1", word)
+    if not word_id:
+        return False
     result = await conn.execute(
         'DELETE FROM word_group_assignments WHERE group_id = $1 AND word_id = $2',
         group_id, word_id
